refactor: move unpackedImageSize to imgutil
